requirements.txt
Django==3.1.4
djangorestframework~=3.11.0
django-cors-headers~=3.2.1
django-environ~=0.4.5
gunicorn~=20.0.4
pandas